# Unidad 5/SpringThymeLeaf/SpringBootThymeleaf/SpringBootThymeleaf-openapi.yaml
openapi: "3.1.0"
info:
  title: "SpringBootThymeleaf API"
  description: "SpringBootThymeleaf API"
  version: "1.0.0"
servers:
  - url: "https://SpringBootThymeleaf"
paths:
...
  /verempleados:
    get:
      summary: "GET verempleados"
      operationId: "mostrarEmpleados"
      parameters:
        - name: "depno"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
  /verempleado:
    get:
      summary: "GET verempleado"
      operationId: "mostrarEmpleado"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"